sdf
df

+++ database/migrations/2022_11_14_201414_create_order_attributes_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('color');
             $table->string('size');
             $table->boolean('status');
-            $table->decimal('amount_paid')->nullable();
+            $table->integer('amount_paid')->nullable();
             $table->foreignIdFor(Order::class)->cascadeOnDelete();
         });
     }
